+++ src/pages/ReservationProPage.js
@@ -1,6 +1,6 @@
 import "../assets/css/common.css";
 
-import { useEffect, useRef } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useSearchParams, useParams } from 'react-router-dom';
 import { getProductsAndSeatInfo, saveReservation } from '../api/reservation/ReservationApi';
@@ -10,11 +10,18 @@ import ProductCompo from "../component/reservation/ProductCompo";
 import CproductCompo from "../component/reservation/CproductCompo";
 import TimeTableCompo from "../component/reservation/TimeTableCompo";
 import { consoleLog } from "../assets/js/jslib";
+import { Swiper, SwiperSlide } from "swiper/react"; // basic
+import SwiperCore, { Navigation, Pagination } from "swiper";
+import "swiper/css"; //basic
+import "swiper/css/navigation";
+import "swiper/css/pagination";
 /**
  * 예약 시도 페이지 
  * @returns page
  */
 function ReservationProPage() {
+  const [slideCount, setSlideCount] = useState(6);
+  SwiperCore.use([Navigation, Pagination]);
   const params = useParams();
   const [searchParams, setSearchParams] = useSearchParams();
   const paymentRef = useRef();
@@ -56,6 +63,30 @@ function ReservationProPage() {
       alert('예약에 실패했습니다');
     }
   }
+  /**
+  * 수량변경시
+  * @param {object} params 
+  */
+  function changeCount(params) {
+    let e = params.event;
+    dispatch(ReserAction.changeCount({ productId: params.productId, count: e.target.value }));
+  }
+  /**
+   * 상품 취소시
+   * @param {int} productId 
+   */
+  function cancel(productId) {
+    dispatch(ReserAction.removeProduct({ productId: productId }));
+  }
+  let windowResize = () => {
+    if (window.innerWidth <= 555) {
+      setSlideCount(3);
+    }
+    setSlideCount(6);
+  };
+  useEffect(() => {
+    window.addEventListener("resize", windowResize);
+}, []);
   return (
     <div>
       <hr></hr>
@@ -65,7 +96,20 @@ function ReservationProPage() {
       <ProductCompo />
       <hr></hr>
       <h2>선택한 상품</h2>
-      <CproductCompo />
+      <Swiper slidesPerView={slideCount}>
+        {state.ReserReducers.choiceProducts.map(product => {
+          return (
+            <SwiperSlide key={`${product.id}cdiv`} className='c_product_box'>
+              <CproductCompo product={product} />
+              <input className="count_in" key={`${product.id}cn`} min="1" type="number" onChange={(event) => changeCount({ productId: product.id, event: event })} />
+              <button key={`${product.id}cc`} onClick={() => { cancel(product.id) }}>취소</button>
+            </SwiperSlide>
+          );
+        })}
+      </Swiper>
+      <div>
+        <p>{state.ReserReducers.totalPrice.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",")}원</p>
+      </div>
       <hr></hr>
       <PaymentCompo ref={paymentRef} />
       <button onClick={order}>예약 하기</button>
